Split directory and file parser classes into separate files
